Moved UnauthorizedUserException to an appropriate folder
