+++ src/main/java/ca/utoronto/fitbook/adapter/web/FollowController.java
@@ -1,6 +1,5 @@
 package ca.utoronto.fitbook.adapter.web;
 
-import ca.utoronto.fitbook.adapter.persistence.UnauthorizedUserException;
 import ca.utoronto.fitbook.application.port.in.FollowUseCase;
 import ca.utoronto.fitbook.application.port.in.command.FollowCommand;
 import ca.utoronto.fitbook.application.port.out.response.FollowResponse;
